Basic h-entry markup.
Adds a custom layout for posts that handles basic h-entry markup. This
pattern of matching an Archetype with a custom Layout is how I plan to
handle other indieweb entry types, like comments, likes, or checkins.
Next steps are to add shortcodes for the markup that I want to handle
dynamically (with otherwise unstyled spans) within content.